+++ src/components/index.js
@@ -1,12 +1,18 @@
-export { default as Anchors } from './anchors'
-export { default as Category } from './parent-category'
-export { default as Emoji } from './parent-emoji'
-export { default as Picker } from './parent-picker'
-export { default as Preview } from './parent-preview'
-export { default as Search } from './parent-search'
-export { default as Skins } from './skins'
-export { default as NimbleCategory } from './category'
-export { default as NimbleEmoji } from './emoji'
-export { default as NimblePicker } from './picker'
-export { default as NimblePreview } from './preview'
-export { default as NimbleSearch } from './search'
+export { default as Anchors } from './anchors/anchors'
+
+export { default as Category } from './category/category'
+export { default as NimbleCategory } from './category/nimble-category'
+
+export { default as Emoji } from './emoji/emoji'
+export { default as NimbleEmoji } from './emoji/nimble-emoji'
+
+export { default as Picker } from './picker/picker'
+export { default as NimblePicker } from './picker/nimble-picker'
+
+export { default as Preview } from './preview/preview'
+export { default as NimblePreview } from './preview/nimble-preview'
+
+export { default as Search } from './search/search'
+export { default as NimbleSearch } from './search/nimble-search'
+
+export { default as Skins } from './skins/skins'
