+++ src/components/picker/picker.js
@@ -1,9 +1,9 @@
-import '../vendor/raf-polyfill'
-
 import React from 'react'
-import data from '../data'
-import { PickerPropTypes, PickerDefaultProps } from '../utils/shared-props'
-import NimblePicker from './picker'
+
+import data from '../../data'
+import NimblePicker from './nimble-picker'
+
+import { PickerPropTypes, PickerDefaultProps } from '../../utils/shared-props'
 
 export default class Picker extends React.PureComponent {
   render() {
