+++ src/components/search/nimble-search.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import NimbleEmojiIndex from '../utils/emoji-index'
-import { SearchPropTypes, SearchDefaultProps } from '../utils/shared-props'
+
+import NimbleEmojiIndex from '../../utils/emoji-index'
+import { SearchPropTypes, SearchDefaultProps } from '../../utils/shared-props'
 
 export default class NimbleSearch extends React.PureComponent {
   constructor(props) {
