+++ src/components/search/search.js
@@ -1,7 +1,9 @@
 import React from 'react'
-import data from '../data'
-import { SearchPropTypes, SearchDefaultProps } from '../utils/shared-props'
-import NimbleSearch from './search'
+
+import data from '../../data'
+import NimbleSearch from './nimble-search'
+
+import { SearchPropTypes, SearchDefaultProps } from '../../utils/shared-props'
 
 export default class Search extends React.PureComponent {
   render() {
